Resolve Issue #23
Now ACCRuntime will connect to the Manager of each worker to release broadcast blocks. Please make sure that Manager is able to accept the request remotely by using IP address.

Future:

User should be able to config the port that Manager uses. This information cannot be fetched from Spark. Now we just assume that all Managers are using the port 1027.