Support MapPartitions
Now we enable a new useful method -- MapParitions.

To use MapPartitions, the developer has to write a call method in the extened Accelerator class as well. However the protocol of the call method is:

def call(in: Iterator[T]): Iterator[U]

instead of the original one.

Also please notice that since we have two protocols of the call method, the developer has to "override" the method instead of defining it directly.

Please refer examples/test/src/main/scala/testapp/TestApp.scala for implementation details.