Add an attribute "isAcc" to the Spark task to indicate if the task may want to use the accelerator.
We whitelist "ShellRDD" as an accelerable RDD. While we build up the RDD
lineage, we propagate the "accelerable" property to the last RDD in a
stage; the last RDD in the stage is associated with a Spark task.
With these changes we can therefore check if a
task is accelerable by simply examing its RDD "accelerable" property.